<refact> : 버튼 아토믹 패턴으로 코드 분리
